Remove abstraction
